+++ src/utils.ts
@@ -48,8 +48,8 @@ export const parseManifest = async (
           'Failed to parse the manifest, please use a valid extension.'
         )
         .then(
-          () => { }, // eslint-disable-line @typescript-eslint/no-empty-function
-          () => { } // eslint-disable-line @typescript-eslint/no-empty-function
+          () => {}, // eslint-disable-line @typescript-eslint/no-empty-function
+          () => {} // eslint-disable-line @typescript-eslint/no-empty-function
         )
       break
   }
@@ -116,8 +116,8 @@ export const execTask = async (
 ): Promise<void> => {
   if (message) {
     window.showInformationMessage(message).then(
-      () => { }, // eslint-disable-line @typescript-eslint/no-empty-function
-      () => { } // eslint-disable-line @typescript-eslint/no-empty-function
+      () => {}, // eslint-disable-line @typescript-eslint/no-empty-function
+      () => {} // eslint-disable-line @typescript-eslint/no-empty-function
     )
   }
   const task = await getTask(mode)
@@ -127,8 +127,8 @@ export const execTask = async (
 
 export const setContext = (ctx: string, state: boolean | string): void => {
   commands.executeCommand('setContext', ctx, state).then(
-    () => { }, // eslint-disable-line @typescript-eslint/no-empty-function
-    () => { } // eslint-disable-line @typescript-eslint/no-empty-function
+    () => {}, // eslint-disable-line @typescript-eslint/no-empty-function
+    () => {} // eslint-disable-line @typescript-eslint/no-empty-function
   )
 }